RAG_chatbot/requirements.txt
langchain
langchain_community
langchain_core
langchain_openai
milvus
pymilvus
python-dotenv